+++ legend-engine-xts-persistence/legend-engine-xt-persistence-component/legend-engine-xt-persistence-component-relational-core/src/main/java/org/finos/legend/engine/persistence/components/relational/api/optimizers/MultiDatasetIngestModeOptimizer.java
@@ -1,0 +1,93 @@
+// Copyright 2025 Goldman Sachs
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package org.finos.legend.engine.persistence.components.relational.api.optimizers;
+
+import org.finos.legend.engine.persistence.components.common.Datasets;
+import org.finos.legend.engine.persistence.components.ingestmode.AppendOnlyAbstract;
+import org.finos.legend.engine.persistence.components.ingestmode.BitemporalDeltaAbstract;
+import org.finos.legend.engine.persistence.components.ingestmode.BitemporalSnapshotAbstract;
+import org.finos.legend.engine.persistence.components.ingestmode.BulkLoadAbstract;
+import org.finos.legend.engine.persistence.components.ingestmode.IngestMode;
+import org.finos.legend.engine.persistence.components.ingestmode.IngestModeVisitor;
+import org.finos.legend.engine.persistence.components.ingestmode.NoOpAbstract;
+import org.finos.legend.engine.persistence.components.ingestmode.NontemporalDeltaAbstract;
+import org.finos.legend.engine.persistence.components.ingestmode.NontemporalSnapshotAbstract;
+import org.finos.legend.engine.persistence.components.ingestmode.UnitemporalDelta;
+import org.finos.legend.engine.persistence.components.ingestmode.UnitemporalDeltaAbstract;
+import org.finos.legend.engine.persistence.components.ingestmode.UnitemporalSnapshotAbstract;
+
+public class MultiDatasetIngestModeOptimizer implements IngestModeVisitor<IngestMode>
+{
+    private Datasets datasets;
+
+    public MultiDatasetIngestModeOptimizer(Datasets datasets)
+    {
+        this.datasets = datasets;
+    }
+
+    @Override
+    public IngestMode visitAppendOnly(AppendOnlyAbstract appendOnly)
+    {
+        return appendOnly;
+    }
+
+    @Override
+    public IngestMode visitNontemporalSnapshot(NontemporalSnapshotAbstract nontemporalSnapshot)
+    {
+        return nontemporalSnapshot;
+    }
+
+    @Override
+    public IngestMode visitNontemporalDelta(NontemporalDeltaAbstract nontemporalDelta)
+    {
+        return nontemporalDelta;
+    }
+
+    @Override
+    public IngestMode visitUnitemporalSnapshot(UnitemporalSnapshotAbstract unitemporalSnapshot)
+    {
+        return unitemporalSnapshot;
+    }
+
+    @Override
+    public IngestMode visitUnitemporalDelta(UnitemporalDeltaAbstract unitemporalDelta)
+    {
+        return new UnitemporalDeltaOptimizer((UnitemporalDelta) unitemporalDelta, datasets).optimize();
+    }
+
+    @Override
+    public IngestMode visitBitemporalSnapshot(BitemporalSnapshotAbstract bitemporalSnapshot)
+    {
+        return bitemporalSnapshot;
+    }
+
+    @Override
+    public IngestMode visitBitemporalDelta(BitemporalDeltaAbstract bitemporalDelta)
+    {
+        return bitemporalDelta;
+    }
+
+    @Override
+    public IngestMode visitBulkLoad(BulkLoadAbstract bulkLoad)
+    {
+        return bulkLoad;
+    }
+
+    @Override
+    public IngestMode visitNoOp(NoOpAbstract noOpAbstract)
+    {
+        return noOpAbstract;
+    }
+}
